moved number converters into own package and remaining common converters one level up
