+++ livingdoc-fixture-converter/src/main/java/org/livingdoc/fixture/converter/number/FloatConverter.kt
@@ -1,4 +1,4 @@
-package org.livingdoc.fixture.converter.common
+package org.livingdoc.fixture.converter.number
 
 open class FloatConverter : AbstractNumberConverter<Float>() {
     override fun convertToTarget(number: Number): Float = number.toFloat()
