+++ livingdoc-fixture-converter/src/main/java/org/livingdoc/fixture/converter/number/IntegerConverter.kt
@@ -1,4 +1,4 @@
-package org.livingdoc.fixture.converter.common
+package org.livingdoc.fixture.converter.number
 
 
 open class IntegerConverter : AbstractNumberConverter<Int>() {
