+++ livingdoc-fixture-converter/src/test/java/org/livingdoc/fixture/converter/StringConverterTest.kt
@@ -1,7 +1,6 @@
-package org.livingdoc.fixture.converter.common
+package org.livingdoc.fixture.converter
 
 import org.assertj.core.api.Assertions.assertThat
-
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
