+++ livingdoc-fixture-converter/src/test/java/org/livingdoc/fixture/converter/number/NumberConverterContract.kt
@@ -1,4 +1,4 @@
-package org.livingdoc.fixture.converter.common
+package org.livingdoc.fixture.converter.number
 
 import com.nhaarman.mockito_kotlin.mock
 import org.assertj.core.api.Assertions.assertThat
